requirements.txt
PyNaCl
discord.py
pandas
dnspython==1.16.0
async-timeout==3.0.1
pillow
pyfiglet
datetime